fix: Fix event timestamp when none provided
