--- app/jobs/create_event_job.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-class CreateEventJob < ApplicationJob
-  queue_as :default
-
-  def perform(organization, params, timestamp, metadata)
-    result = EventsService.new.create(
-      organization: organization,
-      params: params,
-      timestamp: timestamp,
-      metadata: metadata,
-    )
-
-    result.throw_error unless result.success?
-  end
-end
